Module smartcore::api

Common Interfaces and API

This module provides interfaces and uniform API with simple conventions that are used in other modules for supervised and unsupervised learning.

Traits

NoParameters

empty parameters for an estimator, see BiasedEstimator

Predictor

Implements method predict that estimates target value from new data

PredictorBorrow

Implements method predict that estimates target value from new data, with borrowing

SupervisedEstimator

An estimator for supervised learning, that provides method fit to learn from data and training values

SupervisedEstimatorBorrow

An estimator for supervised learning. In this one parameters are borrowed instead of moved, this is useful for parameters that carry references. Also to be used when there is no predictor attached to the estimator.

Transformer

Implements method transform that filters or modifies input data

UnsupervisedEstimator

An estimator for unsupervised learning, that provides method fit to learn from data