Trait smartcore::api::PredictorBorrow

pub trait PredictorBorrow<'a, X, T> {
    fn predict(&self, x: &'a X) -> Result<Vec<T>, Failed>;
}

Implements method predict that estimates target value from new data, with borrowing

Required Methods

fn predict(&self, x: &'a X) -> Result<Vec<T>, Failed>

Estimate target values from new data.

x - NxM matrix with N observations and M features in each observation.