Trait smartcore::api::SupervisedEstimator

pub trait SupervisedEstimator<X, Y, P> {
    fn fit(x: &X, y: &Y, parameters: P) -> Result<Self, Failed>
    where
        Self: Sized,
        P: Clone;
}

An estimator for supervised learning, , that provides method fit to learn from data and training values

Required Methods

fn fit(x: &X, y: &Y, parameters: P) -> Result<Self, Failed> where
Self: Sized,
P: Clone,

Fit a model to a training dataset, estimate model’s parameters.

x - NxM matrix with N observations and M features in each observation.

y - target training values of size N.

parameters - hyperparameters of an algorithm