Trait smartcore::api::SupervisedEstimatorBorrow

pub trait SupervisedEstimatorBorrow<'a, X, Y, P> {
    fn new() -> Self;
    fn fit(x: &'a X, y: &'a Y, parameters: &'a P) -> Result<Self, Failed>
    where
        Self: Sized,
        P: Clone;
}

An estimator for supervised learning. In this one parameters are borrowed instead of moved, this is useful for parameters that carry references. Also to be used when there is no predictor attached to the estimator.

Required Methods

fn new() -> Self

Empty constructor, instantiate an empty estimator. Object is dropped as soon as fit() is called. used to pass around the correct fit() implementation. by calling ::fit(). mostly used to be used with model_selection::cross_validate(...)

fn fit(x: &'a X, y: &'a Y, parameters: &'a P) -> Result<Self, Failed> where
Self: Sized,
P: Clone,

Fit a model to a training dataset, estimate model’s parameters.

x - NxM matrix with N observations and M features in each observation.

y - target training values of size N.

&parameters - hyperparameters of an algorithm