Trait smartcore::api::Transformer

pub trait Transformer<X> {
    fn transform(&self, x: &X) -> Result<X, Failed>;
}

Implements method transform that filters or modifies input data

Required Methods

fn transform(&self, x: &X) -> Result<X, Failed>

Transform data by modifying or filtering it

x - NxM matrix with N observations and M features in each observation.