Module smartcore::dataset::boston

The Boston Housing Dataset

Number of Instances	Number of Attributes	Missing Values?	Associated Tasks:

506	13	No	Regression

The Boston house-price data is derived from information collected by the U.S. Census Service concerning housing in the area of Boston, MA. The dataset has following attributes:

Predictor	Data Type	Target?

CRIM, per capita crime rate by town	Numerical	No

ZN, proportion of residential land zoned for lots over 25,000 sq.ft.	Numerical	No

INDUS, proportion of non-retail business acres per town.	Numerical	No

CHAS, Charles River dummy variable (1 if tract bounds river; 0 otherwise)	Nominal	No

NOX, nitric oxides concentration (parts per 10 million)	Numerical	No

RM, average number of rooms per dwelling	Numerical	No

AGE, proportion of owner-occupied units built prior to 1940	Numerical	No

DIS, weighted distances to five Boston employment centres	Numerical	No

RAD, index of accessibility to radial highways	Ordinal	No

TAX, full-value property-tax rate per $10,000	Numerical	No

PTRATIO, pupil-teacher ratio by town	Numerical	No

B, 1000(Bk - 0.63)^2 where Bk is the proportion of blacks by town	Numerical	No

LSTAT, % lower status of the population	Numerical	No

MEDV, Median value of owner-occupied homes in $1000’s	Numerical	Yes

Functions

load_dataset

Get dataset