Module smartcore::dataset::breast_cancer

Breast Cancer Wisconsin (Diagnostic) Data Set

Diagnostic Wisconsin Breast Cancer database

Number of Instances	Number of Attributes	Missing Values?	Associated Tasks:

569	30	No	Classification

Breast Cancer was collected by Dr. William H. Wolberg, W. Nick Street and Olvi L. Mangasarian. Features are computed from a digitized image of a fine needle aspirate (FNA) of a breast mass and describe characteristics of the cell nuclei present in the image. The dataset has following attributes:

Predictor	Data Type	Target?

Radius (mean of distances from center to points on the perimeter)	Numerical	No

Texture (standard deviation of gray-scale values)	Numerical	No

Perimeter	Numerical	No

Area	Numerical	No

Smoothness (local variation in radius lengths)	Numerical	No

Compactness (perimeter^2 / area - 1.0)	Numerical	No

Concavity (severity of concave portions of the contour)	Numerical	No

Concave points (number of concave portions of the contour)	Numerical	No

Symmetry	Numerical	No

Fractal dimension (“coastline approximation” - 1)	Numerical	No

Has cancer	Nominal	Yes

The mean, standard error, and “worst” or largest (mean of the three worst/largest values) of these features were computed for each image, resulting in 30 features. For instance, field 0 is Mean Radius, field 10 is Radius SE, field 20 is Worst Radius.

Functions

load_dataset

Get dataset