Trait smartcore::linalg::basic::arrays::MutArray

pub trait MutArray<T: Debug + Display + Copy + Sized, S>: Array<T, S> {

    fn set(&mut self, pos: S, x: T);
    fn iterator_mut<'b>(
        &'b mut self, 
        axis: u8
    ) -> Box<dyn Iterator<Item = &'b mut T> + 'b>;

    fn swap(&mut self, a: S, b: S)
    where
        S: Copy,
    { ... }
    fn div_element_mut(&mut self, pos: S, x: T)
    where
        T: Number,
        S: Copy,
    { ... }
    fn mul_element_mut(&mut self, pos: S, x: T)
    where
        T: Number,
        S: Copy,
    { ... }
    fn add_element_mut(&mut self, pos: S, x: T)
    where
        T: Number,
        S: Copy,
    { ... }
    fn sub_element_mut(&mut self, pos: S, x: T)
    where
        T: Number,
        S: Copy,
    { ... }
    fn sub_scalar_mut(&mut self, x: T)
    where
        T: Number,
    { ... }
    fn add_scalar_mut(&mut self, x: T)
    where
        T: Number,
    { ... }
    fn mul_scalar_mut(&mut self, x: T)
    where
        T: Number,
    { ... }
    fn div_scalar_mut(&mut self, x: T)
    where
        T: Number,
    { ... }
    fn add_mut(&mut self, other: &dyn Array<T, S>)
    where
        T: Number,
        S: Eq,
    { ... }
    fn sub_mut(&mut self, other: &dyn Array<T, S>)
    where
        T: Number,
        S: Eq,
    { ... }
    fn mul_mut(&mut self, other: &dyn Array<T, S>)
    where
        T: Number,
        S: Eq,
    { ... }
    fn div_mut(&mut self, other: &dyn Array<T, S>)
    where
        T: Number,
        S: Eq,
    { ... }

}

Abstract methods for mutable Array

Required Methods

fn set(&mut self, pos: S, x: T)

assign value to a position

fn iterator_mut<'b>(
&'b mut self,
axis: u8
) -> Box<dyn Iterator<Item = &'b mut T> + 'b>

iterate over mutable values

Provided Methods

fn swap(&mut self, a: S, b: S) where
S: Copy,

swap values between positions

fn div_element_mut(&mut self, pos: S, x: T) where
T: Number,
S: Copy,

divide element by a given value

fn mul_element_mut(&mut self, pos: S, x: T) where
T: Number,
S: Copy,

multiply element for a given value

fn add_element_mut(&mut self, pos: S, x: T) where
T: Number,
S: Copy,

add a given value to an element

fn sub_element_mut(&mut self, pos: S, x: T) where
T: Number,
S: Copy,

subtract a given value to an element

fn sub_scalar_mut(&mut self, x: T) where
T: Number,

subtract a given value to all the elements

fn add_scalar_mut(&mut self, x: T) where
T: Number,

add a given value to all the elements

fn mul_scalar_mut(&mut self, x: T) where
T: Number,

multiply a given value to all the elements

fn div_scalar_mut(&mut self, x: T) where
T: Number,

divide a given value to all the elements

fn add_mut(&mut self, other: &dyn Array<T, S>) where
T: Number,
S: Eq,

add values from another array to the values of initial array

fn sub_mut(&mut self, other: &dyn Array<T, S>) where
T: Number,
S: Eq,

subtract values from another array to the values of initial array

fn mul_mut(&mut self, other: &dyn Array<T, S>) where
T: Number,
S: Eq,

multiply values from another array to the values of initial array

fn div_mut(&mut self, other: &dyn Array<T, S>) where
T: Number,
S: Eq,

divide values from another array to the values of initial array