Trait smartcore::linalg::basic::arrays::MutArrayView1

pub trait MutArrayView1<T: Debug + Display + Copy + Sized>: MutArray<T, usize> + ArrayView1<T> {
    fn copy_from(&mut self, other: &dyn Array<T, usize>) { ... }
    fn abs_mut(&mut self)
    where
        T: Number + Signed,
    { ... }
    fn neg_mut(&mut self)
    where
        T: Number + Neg<Output = T>,
    { ... }
    fn pow_mut(&mut self, p: T)
    where
        T: RealNumber,
    { ... }
    fn argsort_mut(&mut self) -> Vec<usize>
    where
        T: Number + PartialOrd,
    { ... }
    fn softmax_mut(&mut self)
    where
        T: RealNumber,
    { ... }
}

Trait for mutable 1D-array

Provided Methods

fn copy_from(&mut self, other: &dyn Array<T, usize>)

copy a mutable view from array

fn abs_mut(&mut self) where
T: Number + Signed,

return a mutable view of absolute values

fn neg_mut(&mut self) where
T: Number + Neg<Output = T>,

return a mutable view of values with opposite sign

fn pow_mut(&mut self, p: T) where
T: RealNumber,

return a mutable view of values at power p

fn argsort_mut(&mut self) -> Vec<usize> where
T: Number + PartialOrd,

return vector of indices for sorted elements

fn softmax_mut(&mut self) where
T: RealNumber,

return softmax values