Trait smartcore::linalg::BaseMatrix

pub trait BaseMatrix<T: RealNumber>: Clone + Debug {
    type RowVector: BaseVector<T> + Clone + Debug;

    fn from_row_vector(vec: Self::RowVector) -> Self;
    fn to_row_vector(self) -> Self::RowVector;
    fn get(&self, row: usize, col: usize) -> T;
    fn get_row_as_vec(&self, row: usize) -> Vec<T>;
    fn get_row(&self, row: usize) -> Self::RowVector;
    fn copy_row_as_vec(&self, row: usize, result: &mut Vec<T>);
    fn get_col_as_vec(&self, col: usize) -> Vec<T>;
    fn copy_col_as_vec(&self, col: usize, result: &mut Vec<T>);
    fn set(&mut self, row: usize, col: usize, x: T);
    fn eye(size: usize) -> Self;
    fn zeros(nrows: usize, ncols: usize) -> Self;
    fn ones(nrows: usize, ncols: usize) -> Self;
    fn fill(nrows: usize, ncols: usize, value: T) -> Self;
    fn shape(&self) -> (usize, usize);
    fn h_stack(&self, other: &Self) -> Self;
    fn v_stack(&self, other: &Self) -> Self;
    fn matmul(&self, other: &Self) -> Self;
    fn dot(&self, other: &Self) -> T;
    fn slice(&self, rows: Range<usize>, cols: Range<usize>) -> Self;
    fn approximate_eq(&self, other: &Self, error: T) -> bool;
    fn add_mut(&mut self, other: &Self) -> &Self;
    fn sub_mut(&mut self, other: &Self) -> &Self;
    fn mul_mut(&mut self, other: &Self) -> &Self;
    fn div_mut(&mut self, other: &Self) -> &Self;
    fn div_element_mut(&mut self, row: usize, col: usize, x: T);
    fn mul_element_mut(&mut self, row: usize, col: usize, x: T);
    fn add_element_mut(&mut self, row: usize, col: usize, x: T);
    fn sub_element_mut(&mut self, row: usize, col: usize, x: T);
    fn add_scalar_mut(&mut self, scalar: T) -> &Self;
    fn sub_scalar_mut(&mut self, scalar: T) -> &Self;
    fn mul_scalar_mut(&mut self, scalar: T) -> &Self;
    fn div_scalar_mut(&mut self, scalar: T) -> &Self;
    fn transpose(&self) -> Self;
    fn rand(nrows: usize, ncols: usize) -> Self;
    fn norm2(&self) -> T;
    fn norm(&self, p: T) -> T;
    fn column_mean(&self) -> Vec<T>;
    fn negative_mut(&mut self);
    fn reshape(&self, nrows: usize, ncols: usize) -> Self;
    fn copy_from(&mut self, other: &Self);
    fn abs_mut(&mut self) -> &Self;
    fn sum(&self) -> T;
    fn max(&self) -> T;
    fn min(&self) -> T;
    fn softmax_mut(&mut self);
    fn pow_mut(&mut self, p: T) -> &Self;
    fn argmax(&self) -> Vec<usize>;
    fn unique(&self) -> Vec<T>;
    fn cov(&self) -> Self;

    fn add(&self, other: &Self) -> Self { ... }
    fn sub(&self, other: &Self) -> Self { ... }
    fn mul(&self, other: &Self) -> Self { ... }
    fn div(&self, other: &Self) -> Self { ... }
    fn add_scalar(&self, scalar: T) -> Self { ... }
    fn sub_scalar(&self, scalar: T) -> Self { ... }
    fn mul_scalar(&self, scalar: T) -> Self { ... }
    fn div_scalar(&self, scalar: T) -> Self { ... }
    fn negative(&self) -> Self { ... }
    fn abs(&self) -> Self { ... }
    fn max_diff(&self, other: &Self) -> T { ... }
    fn pow(&mut self, p: T) -> Self { ... }
    fn take(&self, index: &[usize], axis: u8) -> Self { ... }

}

Generic matrix type.

Required Associated Types

type RowVector: BaseVector<T> + Clone + Debug

Row vector that is associated with this matrix type, e.g. if we have an implementation of sparce matrix we should have an associated sparce vector type that represents a row in this matrix.

Required Methods

fn from_row_vector(vec: Self::RowVector) -> Self

Transforms row vector vec into a 1xM matrix.

fn to_row_vector(self) -> Self::RowVector

Transforms 1-d matrix of 1xM into a row vector.

fn get(&self, row: usize, col: usize) -> T

Get an element of the matrix.

row - row number

col - column number

fn get_row_as_vec(&self, row: usize) -> Vec<T>

Get a vector with elements of the row’th row

row - row number

fn get_row(&self, row: usize) -> Self::RowVector

Get the row’th row

row - row number

fn copy_row_as_vec(&self, row: usize, result: &mut Vec<T>)

Copies a vector with elements of the row’th row into result

row - row number

result - receiver for the row

fn get_col_as_vec(&self, col: usize) -> Vec<T>

Get a vector with elements of the col’th column

col - column number

fn copy_col_as_vec(&self, col: usize, result: &mut Vec<T>)

Copies a vector with elements of the col’th column into result

col - column number

result - receiver for the col

fn set(&mut self, row: usize, col: usize, x: T)

Set an element at col, row to x

fn eye(size: usize) -> Self

Create an identity matrix of size size

fn zeros(nrows: usize, ncols: usize) -> Self

Create new matrix with zeros of size nrows by ncols.

fn ones(nrows: usize, ncols: usize) -> Self

Create new matrix with ones of size nrows by ncols.

fn fill(nrows: usize, ncols: usize, value: T) -> Self

Create new matrix of size nrows by ncols where each element is set to value.

fn shape(&self) -> (usize, usize)

Return the shape of an array.

fn h_stack(&self, other: &Self) -> Self

Stack arrays in sequence vertically (row wise).

use smartcore::linalg::naive::dense_matrix::*;

let a = DenseMatrix::from_2d_array(&[&[1., 2., 3.], &[4., 5., 6.]]);
let b = DenseMatrix::from_2d_array(&[&[1., 2.], &[3., 4.]]);
let expected = DenseMatrix::from_2d_array(&[
    &[1., 2., 3., 1., 2.],
    &[4., 5., 6., 3., 4.]
]);

assert_eq!(a.h_stack(&b), expected);

fn v_stack(&self, other: &Self) -> Self

Stack arrays in sequence horizontally (column wise).

use smartcore::linalg::naive::dense_matrix::*;

let a = DenseMatrix::from_array(1, 3, &[1., 2., 3.]);
let b = DenseMatrix::from_array(1, 3, &[4., 5., 6.]);
let expected = DenseMatrix::from_2d_array(&[
    &[1., 2., 3.],
    &[4., 5., 6.]
]);

assert_eq!(a.v_stack(&b), expected);

fn matmul(&self, other: &Self) -> Self

Matrix product.

use smartcore::linalg::naive::dense_matrix::*;

let a = DenseMatrix::from_2d_array(&[&[1., 2.], &[3., 4.]]);
let expected = DenseMatrix::from_2d_array(&[
    &[7., 10.],
    &[15., 22.]
]);

assert_eq!(a.matmul(&a), expected);

fn dot(&self, other: &Self) -> T

Vector dot product Both matrices should be of size 1xM

use smartcore::linalg::naive::dense_matrix::*;

let a = DenseMatrix::from_array(1, 3, &[1., 2., 3.]);
let b = DenseMatrix::from_array(1, 3, &[4., 5., 6.]);

assert_eq!(a.dot(&b), 32.);

fn slice(&self, rows: Range<usize>, cols: Range<usize>) -> Self

Return a slice of the matrix.

rows - range of rows to return

cols - range of columns to return

use smartcore::linalg::naive::dense_matrix::*;

let m = DenseMatrix::from_2d_array(&[
            &[1., 2., 3., 1.],
            &[4., 5., 6., 3.],
            &[7., 8., 9., 5.]
        ]);
let expected = DenseMatrix::from_2d_array(&[&[2., 3.], &[5., 6.]]);
let result = m.slice(0..2, 1..3);
assert_eq!(result, expected);

fn approximate_eq(&self, other: &Self, error: T) -> bool

Returns True if matrices are element-wise equal within a tolerance error.

fn add_mut(&mut self, other: &Self) -> &Self

Add matrices, element-wise, overriding original matrix with result.

fn sub_mut(&mut self, other: &Self) -> &Self

Subtract matrices, element-wise, overriding original matrix with result.

fn mul_mut(&mut self, other: &Self) -> &Self

Multiply matrices, element-wise, overriding original matrix with result.

fn div_mut(&mut self, other: &Self) -> &Self

Divide matrices, element-wise, overriding original matrix with result.

fn div_element_mut(&mut self, row: usize, col: usize, x: T)

Divide single element of the matrix by x, write result to original matrix.

fn mul_element_mut(&mut self, row: usize, col: usize, x: T)

Multiply single element of the matrix by x, write result to original matrix.

fn add_element_mut(&mut self, row: usize, col: usize, x: T)

Add single element of the matrix to x, write result to original matrix.

fn sub_element_mut(&mut self, row: usize, col: usize, x: T)

Subtract x from single element of the matrix, write result to original matrix.

fn add_scalar_mut(&mut self, scalar: T) -> &Self

Add scalar to the matrix, override original matrix with result.

fn sub_scalar_mut(&mut self, scalar: T) -> &Self

Subtract scalar from the elements of matrix, override original matrix with result.

fn mul_scalar_mut(&mut self, scalar: T) -> &Self

Multiply scalar by the elements of matrix, override original matrix with result.

fn div_scalar_mut(&mut self, scalar: T) -> &Self

Divide elements of the matrix by scalar, override original matrix with result.

fn transpose(&self) -> Self

Reverse or permute the axes of the matrix, return new matrix.

fn rand(nrows: usize, ncols: usize) -> Self

Create new nrows by ncols matrix and populate it with random samples from a uniform distribution over [0, 1).

fn norm2(&self) -> T

Returns L2 norm.

fn norm(&self, p: T) -> T

Returns matrix norm of order p.

fn column_mean(&self) -> Vec<T>

Returns the average of the matrix columns.

fn negative_mut(&mut self)

Numerical negative, element-wise. Overrides original matrix.

fn reshape(&self, nrows: usize, ncols: usize) -> Self

Returns new matrix of shape nrows by ncols with data copied from original matrix.

use smartcore::linalg::naive::dense_matrix::*;

let a = DenseMatrix::from_array(1, 6, &[1., 2., 3., 4., 5., 6.]);
let expected = DenseMatrix::from_2d_array(&[
            &[1., 2., 3.],
            &[4., 5., 6.]
        ]);

assert_eq!(a.reshape(2, 3), expected);

fn copy_from(&mut self, other: &Self)

Copies content of other matrix.

fn abs_mut(&mut self) -> &Self

Calculate the absolute value element-wise. Overrides original matrix.

fn sum(&self) -> T

Calculates sum of all elements of the matrix.

fn max(&self) -> T

Calculates max of all elements of the matrix.

fn min(&self) -> T

Calculates min of all elements of the matrix.

fn softmax_mut(&mut self)

Calculates Softmax function. Overrides the matrix with result.

fn pow_mut(&mut self, p: T) -> &Self

Raises elements of the matrix to the power of p

fn argmax(&self) -> Vec<usize>

Returns the indices of the maximum values in each row.

use smartcore::linalg::naive::dense_matrix::*;
let a = DenseMatrix::from_array(2, 3, &[1., 2., 3., -5., -6., -7.]);

assert_eq!(a.argmax(), vec![2, 0]);

fn unique(&self) -> Vec<T>

Returns vector with unique values from the matrix.

use smartcore::linalg::naive::dense_matrix::*;
let a = DenseMatrix::from_array(3, 3, &[1., 2., 2., -2., -6., -7., 2., 3., 4.]);

assert_eq!(a.unique(), vec![-7., -6., -2., 1., 2., 3., 4.]);

fn cov(&self) -> Self

Calculates the covariance matrix

Provided Methods

fn add(&self, other: &Self) -> Self

Add matrices, element-wise

fn sub(&self, other: &Self) -> Self

Subtract matrices, element-wise

fn mul(&self, other: &Self) -> Self

Multiply matrices, element-wise

fn div(&self, other: &Self) -> Self

Divide matrices, element-wise

fn add_scalar(&self, scalar: T) -> Self

Add scalar to the matrix.

fn sub_scalar(&self, scalar: T) -> Self

Subtract scalar from the elements of matrix.

fn mul_scalar(&self, scalar: T) -> Self

Multiply scalar by the elements of matrix.

fn div_scalar(&self, scalar: T) -> Self

Divide elements of the matrix by scalar.

fn negative(&self) -> Self

Numerical negative, element-wise.

fn abs(&self) -> Self

Calculate the absolute value element-wise.

fn max_diff(&self, other: &Self) -> T

Calculates max(|a - b|) of two matrices

use smartcore::linalg::naive::dense_matrix::*;

let a = DenseMatrix::from_array(2, 3, &[1., 2., 3., 4., -5., 6.]);
let b = DenseMatrix::from_array(2, 3, &[2., 3., 4., 1., 0., -12.]);

assert_eq!(a.max_diff(&b), 18.);
assert_eq!(b.max_diff(&b), 0.);

fn pow(&mut self, p: T) -> Self

Returns new matrix with elements raised to the power of p

fn take(&self, index: &[usize], axis: u8) -> Self

Take elements from an array along an axis.