Trait smartcore::linalg::BaseVector

pub trait BaseVector<T: RealNumber>: Clone + Debug {

    fn get(&self, i: usize) -> T;
    fn set(&mut self, i: usize, x: T);
    fn len(&self) -> usize;
    fn to_vec(&self) -> Vec<T>;
    fn zeros(len: usize) -> Self;
    fn ones(len: usize) -> Self;
    fn fill(len: usize, value: T) -> Self;
    fn dot(&self, other: &Self) -> T;
    fn approximate_eq(&self, other: &Self, error: T) -> bool;
    fn norm2(&self) -> T;
    fn norm(&self, p: T) -> T;
    fn div_element_mut(&mut self, pos: usize, x: T);
    fn mul_element_mut(&mut self, pos: usize, x: T);
    fn add_element_mut(&mut self, pos: usize, x: T);
    fn sub_element_mut(&mut self, pos: usize, x: T);
    fn add_mut(&mut self, other: &Self) -> &Self;
    fn sub_mut(&mut self, other: &Self) -> &Self;
    fn mul_mut(&mut self, other: &Self) -> &Self;
    fn div_mut(&mut self, other: &Self) -> &Self;
    fn sum(&self) -> T;
    fn unique(&self) -> Vec<T>;
    fn copy_from(&mut self, other: &Self);

    fn is_empty(&self) -> bool { ... }
    fn from_array(f: &[T]) -> Self { ... }
    fn sub_scalar_mut(&mut self, x: T) -> &Self { ... }
    fn add_scalar_mut(&mut self, x: T) -> &Self { ... }
    fn mul_scalar_mut(&mut self, x: T) -> &Self { ... }
    fn div_scalar_mut(&mut self, x: T) -> &Self { ... }
    fn add_scalar(&self, x: T) -> Self { ... }
    fn sub_scalar(&self, x: T) -> Self { ... }
    fn mul_scalar(&self, x: T) -> Self { ... }
    fn div_scalar(&self, x: T) -> Self { ... }
    fn add(&self, other: &Self) -> Self { ... }
    fn sub(&self, other: &Self) -> Self { ... }
    fn mul(&self, other: &Self) -> Self { ... }
    fn div(&self, other: &Self) -> Self { ... }
    fn mean(&self) -> T { ... }
    fn var(&self) -> T { ... }
    fn std(&self) -> T { ... }
    fn take(&self, index: &[usize]) -> Self { ... }

}

Column or row vector

Required Methods

fn get(&self, i: usize) -> T

Get an element of a vector

i - index of an element

fn set(&mut self, i: usize, x: T)

Set an element at i to x

i - index of an element

x - new value

fn len(&self) -> usize

Get number of elevemnt in the vector

fn to_vec(&self) -> Vec<T>

Return a vector with the elements of the one-dimensional array.

fn zeros(len: usize) -> Self

Create new vector with zeros of size len.

fn ones(len: usize) -> Self

Create new vector with ones of size len.

fn fill(len: usize, value: T) -> Self

Create new vector of size len where each element is set to value.

fn dot(&self, other: &Self) -> T

Vector dot product

fn approximate_eq(&self, other: &Self, error: T) -> bool

Returns True if matrices are element-wise equal within a tolerance error.

fn norm2(&self) -> T

Returns [L2 norm] of the vector(https://en.wikipedia.org/wiki/Matrix_norm).

fn norm(&self, p: T) -> T

Returns vectors norm of order p.

fn div_element_mut(&mut self, pos: usize, x: T)

Divide single element of the vector by x, write result to original vector.

fn mul_element_mut(&mut self, pos: usize, x: T)

Multiply single element of the vector by x, write result to original vector.

fn add_element_mut(&mut self, pos: usize, x: T)

Add single element of the vector to x, write result to original vector.

fn sub_element_mut(&mut self, pos: usize, x: T)

Subtract x from single element of the vector, write result to original vector.

fn add_mut(&mut self, other: &Self) -> &Self

Add vectors, element-wise, overriding original vector with result.

fn sub_mut(&mut self, other: &Self) -> &Self

Subtract vectors, element-wise, overriding original vector with result.

fn mul_mut(&mut self, other: &Self) -> &Self

Multiply vectors, element-wise, overriding original vector with result.

fn div_mut(&mut self, other: &Self) -> &Self

Divide vectors, element-wise, overriding original vector with result.

fn sum(&self) -> T

Calculates sum of all elements of the vector.

fn unique(&self) -> Vec<T>

Returns unique values from the vector.

use smartcore::linalg::naive::dense_matrix::*;
let a = vec!(1., 2., 2., -2., -6., -7., 2., 3., 4.);

assert_eq!(a.unique(), vec![-7., -6., -2., 1., 2., 3., 4.]);

fn copy_from(&mut self, other: &Self)

Copies content of other vector.

Provided Methods

fn is_empty(&self) -> bool

Returns true if the vector is empty.

fn from_array(f: &[T]) -> Self

Create a new vector from a &[T]

use smartcore::linalg::naive::dense_matrix::*;
let a: [f64; 5] = [0., 0.5, 2., 3., 4.];
let v: Vec<f64> = BaseVector::from_array(&a);
assert_eq!(v, vec![0., 0.5, 2., 3., 4.]);

fn sub_scalar_mut(&mut self, x: T) -> &Self

Subtract scalar

fn add_scalar_mut(&mut self, x: T) -> &Self

Subtract scalar

fn mul_scalar_mut(&mut self, x: T) -> &Self

Subtract scalar

fn div_scalar_mut(&mut self, x: T) -> &Self

Subtract scalar

fn add_scalar(&self, x: T) -> Self

Add vectors, element-wise

fn sub_scalar(&self, x: T) -> Self

Subtract vectors, element-wise

fn mul_scalar(&self, x: T) -> Self

Multiply vectors, element-wise

fn div_scalar(&self, x: T) -> Self

Divide vectors, element-wise

fn add(&self, other: &Self) -> Self

Add vectors, element-wise

fn sub(&self, other: &Self) -> Self

Subtract vectors, element-wise

fn mul(&self, other: &Self) -> Self

Multiply vectors, element-wise

fn div(&self, other: &Self) -> Self

Divide vectors, element-wise

fn mean(&self) -> T

Computes the arithmetic mean.

fn var(&self) -> T

Computes variance.

fn std(&self) -> T

Computes the standard deviation.

fn take(&self, index: &[usize]) -> Self

Take elements from an array.