Module smartcore::linear

Supervised classification and regression models that assume linear relationship between dependent and explanatory variables.

Linear Models

Linear models describe a continuous response variable as a function of one or more predictor variables. The model describes the relationship between a dependent variable y (also called the response) as a function of one or more independent, or explanatory variables \(X_i\). The general equation for a linear model is: \[y = \beta_0 + \sum_{i=1}^n \beta_iX_i + \epsilon\]

where \(\beta_0 \) is the intercept term (the expected value of Y when X = 0), \(\epsilon \) is an error term that is is independent of X and \(\beta_i \) is the average increase in y associated with a one-unit increase in \(X_i\)

Model assumptions:

Linearity. The relationship between X and the mean of y is linear.

Constant variance. The variance of residual is the same for any value of X.

Normality. For any fixed value of X, Y is normally distributed.

Independence. Observations are independent of each other.

References:

“An Introduction to Statistical Learning”, James G., Witten D., Hastie T., Tibshirani R., 3. Linear Regression

“The Statistical Sleuth, A Course in Methods of Data Analysis”, Ramsey F.L., Schafer D.W., Ch 7, 8, 3rd edition, 2013

Modules

elastic_net

Elastic Net

lasso

Lasso

linear_regression

Linear Regression

logistic_regression

Logistic Regression

ridge_regression

Ridge Regression