Function smartcore::model_selection::cross_validate

pub fn cross_validate<TX, TY, X, Y, H, E, K, S>(
    _estimator: E, 
    x: &X, 
    y: &Y, 
    parameters: H, 
    cv: &K, 
    score: &S
) -> Result<CrossValidationResult, Failed> where
    TX: Number + RealNumber,
    TY: Number,
    X: Array2<TX>,
    Y: Array1<TY>,
    H: Clone,
    K: BaseKFold,
    E: SupervisedEstimator<X, Y, H>,
    S: Fn(&Y, &Y) -> f64, 

Evaluate an estimator by cross-validation using given metric.

fit_estimator - a fit function of an estimator

x - features, matrix of size NxM where N is number of samples and M is number of attributes.

y - target values, should be of size N

parameters - parameters of selected estimator. Use Default::default() for default parameters.

cv - the cross-validation splitting strategy, should be an instance of BaseKFold

score - a metric to use for evaluation, see metrics