Trait smartcore::model_selection::BaseKFold

pub trait BaseKFold {
    type Output: Iterator<Item = (Vec<usize>, Vec<usize>)>;

    fn split<T: RealNumber, M: Matrix<T>>(&self, x: &M) -> Self::Output;
    fn n_splits(&self) -> usize;
}

An interface for the K-Folds cross-validator

Required Associated Types

type Output: Iterator<Item = (Vec<usize>, Vec<usize>)>

An iterator over indices that split data into training and test set.

Required Methods

fn split<T: RealNumber, M: Matrix<T>>(&self, x: &M) -> Self::Output

Return a tuple containing the the training set indices for that split and the testing set indices for that split.

fn n_splits(&self) -> usize

Returns the number of splits