Module smartcore::preprocessing::categorical

Transform a data matrix by replaceing all categorical variables with their one-hot vector equivalents

One-hot Encoding For RealNumber Matricies

Transform a data Matrix by replacing all categorical variables with their one-hot equivalents

Internally OneHotEncoder treats every categorical column as a series and transforms it using CategoryMapper

Usage Example

use smartcore::linalg::naive::dense_matrix::DenseMatrix;
use smartcore::preprocessing::categorical::{OneHotEncoder, OneHotEncoderParams};
let data = DenseMatrix::from_2d_array(&[
        &[1.5, 1.0, 1.5, 3.0],
        &[1.5, 2.0, 1.5, 4.0],
        &[1.5, 1.0, 1.5, 5.0],
        &[1.5, 2.0, 1.5, 6.0],
  ]);
let encoder_params = OneHotEncoderParams::from_cat_idx(&[1, 3]);
// Infer number of categories from data and return a reusable encoder
let encoder = OneHotEncoder::fit(&data, encoder_params).unwrap();
// Transform categorical to one-hot encoded (can transform similar)
let oh_data = encoder.transform(&data).unwrap();
// Produces the following:
//    &[1.5, 1.0, 0.0, 1.5, 1.0, 0.0, 0.0, 0.0]
//    &[1.5, 0.0, 1.0, 1.5, 0.0, 1.0, 0.0, 0.0]
//    &[1.5, 1.0, 0.0, 1.5, 0.0, 0.0, 1.0, 0.0]
//    &[1.5, 0.0, 1.0, 1.5, 0.0, 0.0, 0.0, 1.0]

Structs

OneHotEncoder

Encode Categorical variavbles of data matrix to one-hot

OneHotEncoderParams

OneHotEncoder Parameters