Module smartcore::preprocessing::numerical

Preprocess numerical matrices.

Standard-Scaling For RealNumber Matricies

Transform a data Matrix by removing the mean and scaling to unit variance.

Usage Example

use smartcore::api::{Transformer, UnsupervisedEstimator};
use smartcore::linalg::basic::matrix::DenseMatrix;
use smartcore::preprocessing::numerical;
let data = DenseMatrix::from_2d_vec(&vec![
    vec![0.0, 0.0],
    vec![0.0, 0.0],
    vec![1.0, 1.0],
    vec![1.0, 1.0],
]);

let standard_scaler =
numerical::StandardScaler::fit(&data, numerical::StandardScalerParameters::default())
   .unwrap();
let transformed_data = standard_scaler.transform(&data).unwrap();
assert_eq!(
    transformed_data,
    DenseMatrix::from_2d_vec(&vec![
        vec![-1.0, -1.0],
        vec![-1.0, -1.0],
        vec![1.0, 1.0],
        vec![1.0, 1.0],
    ])
);

Structs

StandardScaler

With the StandardScaler data can be adjusted so that every column has a mean of zero and a standard deviation of one. This can improve model training for scaling sensitive models like neural network or nearest neighbors based models.

StandardScalerParameters

Configure Behaviour of StandardScaler.