Trait smartcore::api::Predictor

pub trait Predictor<X, Y> {
    fn predict(&self, x: &X) -> Result<Y, Failed>;
}

Implements method predict that estimates target value from new data

Required Methods

fn predict(&self, x: &X) -> Result<Y, Failed>

Estimate target values from new data.

x - NxM matrix with N observations and M features in each observation.