Trait smartcore::api::SupervisedEstimator

pub trait SupervisedEstimator<X, Y, P>: Predictor<X, Y> {
    fn new() -> Self;
    fn fit(x: &X, y: &Y, parameters: P) -> Result<Self, Failed>
    where
        Self: Sized,
        P: Clone;
}

An estimator for supervised learning, that provides method fit to learn from data and training values

Required Methods

fn new() -> Self

Empty constructor, instantiate an empty estimator. Object is dropped as soon as fit() is called. used to pass around the correct fit() implementation. by calling ::fit(). mostly used to be used with model_selection::cross_validate(...)

fn fit(x: &X, y: &Y, parameters: P) -> Result<Self, Failed> where
Self: Sized,
P: Clone,

Fit a model to a training dataset, estimate model’s parameters.

x - NxM matrix with N observations and M features in each observation.

y - target training values of size N.

parameters - hyperparameters of an algorithm