Trait smartcore::api::UnsupervisedEstimator

pub trait UnsupervisedEstimator<X, P> {
    fn fit(x: &X, parameters: P) -> Result<Self, Failed>
    where
        Self: Sized,
        P: Clone;
}

An estimator for unsupervised learning, that provides method fit to learn from data

Required Methods

fn fit(x: &X, parameters: P) -> Result<Self, Failed> where
Self: Sized,
P: Clone,

Fit a model to a training dataset, estimate model’s parameters.

x - NxM matrix with N observations and M features in each observation.

parameters - hyperparameters of an algorithm